Don’t start the server in CI

+++ test/run.sh
@@ -3,10 +3,13 @@
 echo "Setting up DB"
 mongo test/setup/setupdb.js
 
-echo "Starting test server"
-node app.js --server:port 3102 --database:name postest &
-TESTPID=$(echo $!)
-echo "Test server started with pid $TESTPID"
+
+if [ ! -z "$TRAVIS" ]; then
+    echo "Starting test server"
+    node app.js --server:port 3102 --database:name postest &
+    TESTPID=$(echo $!)
+    echo "Test server started with pid $TESTPID"
+fi
 
 echo "Running mocha tests"
 ./node_modules/.bin/mocha test/model/*.js
@@ -14,5 +17,7 @@ echo "Running mocha tests"
 echo "Running protractor tests"
 ./node_modules/.bin/protractor test/conf/protractor.js
 
-echo "Killing testserver"
-kill $TESTPID
+if [ ! -z "$TRAVIS" ]; then
+    echo "Killing testserver"
+    kill $TESTPID
+fi
